+++ src/churchsong/__main__.py
@@ -62,7 +62,7 @@ def parse_datetime(date_str: str) -> datetime.datetime:
 
 
 def cmd_self_info(_args: argparse.Namespace, config: Configuration) -> None:
-    sys.stderr.write(f'Application version: {get_app_version(config)}\n')
+    sys.stderr.write(f'Installed version:   {get_app_version(config)}\n')
     if latest := get_latest_version(config):
         sys.stderr.write(f'Latest version:      {latest}\n')
     sys.stderr.write(f'Configuration file:  {config.config_toml}\n')